remove synch old and check api

+++ pandora_console/include/functions_api.php
@@ -2129,6 +2129,14 @@ function api_set_create_os($thrash1, $thrash2, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     $values = [];
 
     $values['name'] = $other['data'][0];
@@ -2158,6 +2166,14 @@ function api_set_update_os($id_os, $thrash2, $other, $thrash3)
 
     if (!check_acl($config['id_user'], 0, 'AW')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5157,12 +5173,20 @@ function api_set_update_snmp_module($id_module, $thrash1, $other, $thrash3)
 function api_set_new_network_component($id, $thrash1, $other, $thrash2)
 {
     global $config;
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
         return;
     }
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5251,12 +5275,20 @@ function api_set_new_plugin_component($id, $thrash1, $other, $thrash2)
 {
     global $config;
 
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
         return;
     }
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5523,7 +5555,15 @@ function api_set_new_local_component($id, $thrash1, $other, $thrash2)
 {
     global $config;
 
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5691,7 +5731,15 @@ function api_set_create_alert_template($name, $thrash1, $other, $thrash3)
 {
     global $config;
 
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5826,7 +5874,15 @@ function api_set_update_alert_template($id_template, $thrash1, $other, $thrash3)
 {
     global $config;
 
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -5970,7 +6026,15 @@ function api_set_delete_alert_template($id_template, $thrash1, $other, $thrash3)
 {
     global $config;
 
-    if (defined('METACONSOLE')) {
+    if (is_metaconsole() === true) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -7160,6 +7224,14 @@ function api_set_tag($id, $thrash1, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     $values = [];
     $values['name'] = $id;
     $values['description'] = $other['data'][0];
@@ -7409,8 +7481,6 @@ function api_set_planned_downtimes_deleted($id, $thrash1, $thrash2, $returnType)
  *                       <type_periodicity>
  * @param $thrash3 Don't use.
  */
-
-
 function api_set_planned_downtimes_created($id, $thrash1, $other, $thrash3)
 {
     global $config;
@@ -7565,6 +7635,14 @@ function api_set_update_data_module_policy($id, $thrash1, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($id == '') {
         returnError(
             'The data module could not be updated in policy. Id_policy cannot be left blank.'
@@ -7674,6 +7752,14 @@ function api_set_update_data_module_policy($id, $thrash1, $other, $thrash3)
 function api_set_add_network_module_policy($id, $thrash1, $other, $thrash3)
 {
     if (defined('METACONSOLE')) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -7792,6 +7878,14 @@ function api_set_update_network_module_policy($id, $thrash1, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($id == '') {
         returnError(
             'The network module could not be updated in policy. Id_policy cannot be left blank.'
@@ -7888,6 +7982,14 @@ function api_set_update_network_module_policy($id, $thrash1, $other, $thrash3)
 function api_set_add_plugin_module_policy($id, $thrash1, $other, $thrash3)
 {
     if (defined('METACONSOLE')) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -7996,6 +8098,14 @@ function api_set_add_plugin_module_policy($id, $thrash1, $other, $thrash3)
 function api_set_update_plugin_module_policy($id, $thrash1, $other, $thrash3)
 {
     if (defined('METACONSOLE')) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -8287,6 +8397,14 @@ function api_set_add_snmp_module_policy($id, $thrash1, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($id == '') {
         returnError('The SNMP module could not be added to policy. Id_policy cannot be left blank.');
         return;
@@ -8443,6 +8561,14 @@ function api_set_add_snmp_module_policy($id, $thrash1, $other, $thrash3)
 function api_set_update_snmp_module_policy($id, $thrash1, $other, $thrash3)
 {
     if (defined('METACONSOLE')) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -8603,6 +8729,14 @@ function api_set_remove_agent_from_policy_by_id($id, $thrash1, $other, $thrash2)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($other['data'][0] == '' || !$other['data'][0]) {
         returnError('The agent could not be deleted from policy. Agent cannot be left blank.');
         return;
@@ -8633,6 +8767,14 @@ function api_set_remove_agent_from_policy_by_name($id, $thrash1, $other, $thrash
 {
     if ($id == '' || !$id) {
         returnError('The agent could not be deleted from policy. Policy cannot be left blank.');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -8669,6 +8811,14 @@ function api_set_create_group($id, $thrash1, $other, $thrash3)
     global $config;
 
     if (is_metaconsole()) {
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -8787,6 +8937,14 @@ function api_set_update_group($id_group, $thrash2, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if (db_get_value('id_grupo', 'tgrupo', 'id_grupo', $id_group) === false) {
         returnError('There is no group with the ID provided');
         return;
@@ -8853,6 +9011,14 @@ function api_set_delete_group($id_group, $thrash2, $other, $thrash3)
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -9173,6 +9339,14 @@ function api_set_new_user($id, $thrash2, $other, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     $values = [];
     $values['fullname'] = $other['data'][0];
     $values['firstname'] = $other['data'][1];
@@ -9220,6 +9394,14 @@ function api_set_update_user($id, $thrash2, $other, $thrash3)
 
     if (!check_acl($config['id_user'], 0, 'UM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -9302,8 +9484,6 @@ function api_set_update_user($id, $thrash2, $other, $thrash3)
  *
  * @param $thrash3 Don't use.
  */
-
-
 function api_set_enable_disable_user($id, $thrash2, $other, $thrash3)
 {
     global $config;
@@ -9314,6 +9494,14 @@ function api_set_enable_disable_user($id, $thrash2, $other, $thrash3)
 
     if (!check_acl($config['id_user'], 0, 'UM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -9574,6 +9762,14 @@ function api_set_new_alert_template($id, $id2, $other, $trash1)
 
     if (!check_acl($config['id_user'], 0, 'LM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -9992,6 +10188,14 @@ function api_set_alert_actions($id, $id2, $other, $trash1)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($other['type'] == 'string') {
         returnError('Parameter error.');
         return;
@@ -10140,6 +10344,14 @@ function api_set_new_module_group($id, $thrash2, $other, $trash1)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
         return;
@@ -10165,55 +10377,6 @@ function api_set_new_module_group($id, $thrash2, $other, $trash1)
         returnData('string', ['type' => 'string', 'data' => $return]);
     }
 
-}
-
-
-/**
- * synchronize module group
- *
- * @param $other as server_names (mandatory)
- example:
-
- * api.php?op=set&op2=module_group_synch&other=server_name1|server_name2|server_name3&other_mode=url_encode_separator_|&apipass=1234&user=admin&pass=pandora
- */
-
-
-function api_set_module_group_synch($thrash1, $thrash2, $other, $thrash4)
-{
-    global $config;
-    enterprise_include_once('meta/include/functions_meta.php');
-
-    if (is_metaconsole()) {
-        if (!check_acl($config['id_user'], 0, 'PM')) {
-            returnError('forbidden', 'string');
-            return;
-        }
-
-        $targets = [];
-        foreach ($other['data'] as $server) {
-            $targets[] = $server;
-        }
-
-        $return = meta_module_group_synchronizing($targets, true);
-
-        $module_group_update_err = $return['module_group_update_err'];
-        $module_group_create_err = $return['module_group_create_err'];
-        $module_group_update_ok = $return['module_group_update_ok'];
-        $module_group_create_ok = $return['module_group_create_ok'];
-
-        $string_ok = __('Created/Updated %s/%s module groups', $module_group_create_ok, $module_group_update_ok);
-
-        // User feedback
-        if ($module_group_create_err > 0 or $module_group_update_err > 0) {
-            returnError(sprintf('The module groups %s/%s could not be created/updated <br>', $module_group_create_err, $module_group_update_err));
-        }
-
-        if ($module_group_create_ok > 0 or $module_group_update_ok > 0) {
-            returnData('string', ['type' => 'string', 'data' => $string_ok]);
-        }
-    } else {
-        returnError('This function is for metaconsole only');
-    }
 }
 
 
@@ -10252,6 +10415,14 @@ function api_set_alert_commands($id, $thrash2, $other, $trash1)
 
     if (!check_acl($config['id_user'], 0, 'LW')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -11527,6 +11698,14 @@ function api_set_delete_user($id, $thrash1, $thrash2, $thrash3)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if (!delete_user($id)) {
         returnError('The user could not be deleted');
     } else {
@@ -11559,6 +11738,14 @@ function api_set_add_user_profile($id, $thrash1, $other, $thrash2)
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -11612,6 +11799,14 @@ function api_set_delete_user_profile($id, $thrash1, $other, $thrash2)
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -11722,6 +11917,14 @@ function api_set_create_user_profile_info($thrash1, $thrash2, $other, $returnTyp
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     $values = [
         'name'                => (string) $other['data'][0],
         'agent_view'          => (bool) $other['data'][1] ? 1 : 0,
@@ -11772,6 +11975,14 @@ function api_set_update_user_profile_info($id_profile, $thrash1, $other, $return
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -11834,6 +12045,14 @@ function api_set_delete_user_profile_info($id_profile, $thrash1, $thrash2, $retu
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     $profile = db_get_value('id_perfil', 'tperfil', 'id_perfil', $id_profile);
     if ($profile === false) {
         returnError('id_not_found', 'string');
@@ -11859,8 +12078,6 @@ function api_set_delete_user_profile_info($id_profile, $thrash1, $thrash2, $retu
  * @param $thrash4 Don't use.
  * // http://localhost/pandora_console/include/api.php?op=set&op2=enable_module&id=garfio&id2=Status
  */
-
-
 function api_set_disable_module($agent_name, $module_name, $other, $thrash4)
 {
     if (defined('METACONSOLE')) {
@@ -12720,6 +12937,14 @@ function api_set_create_tag($id, $trash1, $other, $returnType)
 
     if (!check_acl($config['id_user'], 0, 'PM')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
 
@@ -15930,9 +16155,6 @@ function api_get_user_info($thrash1, $thrash2, $other, $returnType)
     If you mark 1 you will avoid the access to the non-administrators users, returning the response `denied' and registering that expulsion in the audit of pandora fms.
 
 */
-
-
-
 function api_set_access_process($thrash1, $thrash2, $other, $returnType)
 {
     if (defined('METACONSOLE')) {
@@ -16591,6 +16813,14 @@ function api_set_delete_user_permission($thrash1, $thrash2, $other, $returnType)
         return;
     }
 
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
+        return;
+    }
+
     if ($other['data'][0] != '') {
         $values = [
             'id_up' => io_safe_output($other['data'][0]),
@@ -16641,6 +16871,14 @@ function api_set_add_permission_user_to_group($thrash1, $thrash2, $other, $retur
 
     if (!check_acl($config['id_user'], 0, 'AW')) {
         returnError('forbidden', 'string');
+        return;
+    }
+
+    $headers = getallheaders();
+    if (isset($headers['idk']) === false
+        && is_management_allowed($headers['idk']) === false
+    ) {
+        returnError('centralized');
         return;
     }
